+++ backend/open_webui/routers/images.py
@@ -1,3 +1,10 @@
+"""
+Modification Log:
+------------------
+| Date       | Author         | MOD TAG            | Description                                                                                         |
+|------------|----------------|--------------------|-----------------------------------------------------------------------------------------------------|
+| 2025-08-01 | AAK7S          | AMER-ENH-V1.2.2    | Admin activity logs                                                                                 |
+"""
 import asyncio
 import base64
 import io