+++ backend/open_webui/routers/pipelines.py
@@ -1,3 +1,11 @@
+"""
+Modification Log:
+------------------
+| Date       | Author         | MOD TAG            | Description                                                                                         |
+|------------|----------------|--------------------|-----------------------------------------------------------------------------------------------------|
+| 2025-08-01 | AAK7S          | AMER-ENH-V1.2.2    | Admin Activity Logs                                                                                 |
+
+"""
 from fastapi import (
     Depends,
     FastAPI,