+++ src/routes/(app)/+layout.svelte
@@ -59,11 +59,13 @@
 
 	onMount(async () => {
 		if ($user === undefined) {
+                       sessionStorage.setItem('attemptSilentLogin', 'true');
 			await goto('/auth');
 		} else if (['user', 'admin'].includes($user.role)) {
 			try {
 				// Check if IndexedDB exists
-				DB = await openDB('Chats', 1);
+                               DB = await openDB('Chats', 1);
+
 
 				if (DB) {
 					const chats = await DB.getAllFromIndex('chats', 'timestamp');